Allow a logged in user to logout
We have a logout link that is only visible if logged in. When clicked we
simply clear the user id from the session and redirect to the homepage.
I've added a `login` and `logout` method to the `ApplicationController`
to clean up the controller actions.

I'm thinking about moving the `current_user`, `login` and `logout`
methods to a separate class or operation that can manage them together
as they all need the session and are related. This will enable us to
move this logic out of the controllers. I'll come back to that later.