Add initial styles based on Flanzo theme
After purchasing the Flanzo bootstrap theme, this is the inital import
of the required assets and also converting the html to slim. Also
removed various images and html from the original theme that are not
required. The slim and css needs cleaning up as this is just a direct
import.